@@ -1,13 +1,7 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-paths";
-import tsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
-// https://vitejs.dev/config/
+// https://vite.dev/config/
 export default defineConfig({
-    // This changes the out put dir from dist to build
-    // comment this out if that isn't relevant for your project
-    build: {
-        outDir: "build",
-    },
-    plugins: [tsconfigPaths(), react()],
-});
+  plugins: [react()],
+})
